Perforance Fixes
-Buses now Buffer the adjacent IFluidHandler
-Removed unnecessary casts

+++ src/main/java/extracells/blocks/BlockHardMEDrive.java
@@ -19,7 +19,7 @@ import net.minecraftforge.common.ForgeDirection;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import extracells.Extracells;
-import extracells.tile.TileEntityHardMEDrive;
+import extracells.tileentity.TileEntityHardMEDrive;
 
 public class BlockHardMEDrive extends RotatableColorBlock
 {
@@ -87,7 +87,7 @@ public class BlockHardMEDrive extends RotatableColorBlock
 	public void onBlockPlacedBy(World world, int x, int y, int z, EntityLivingBase player, ItemStack itemstack)
 	{
 		super.onBlockPlacedBy(world, x, y, z, player, itemstack);
-		int l = MathHelper.floor_double((double) (player.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+		int l = MathHelper.floor_double(player.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
 
 		if (!player.isSneaking())
 		{
